requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-generator==1.10
async-timeout==4.0.2
attrs==23.1.0
cachetools==5.3.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
chromedriver-autoinstaller==0.4.0
colorama==0.4.6
dnspython==2.3.0
exceptiongroup==1.1.1
frozenlist==1.3.3
google-auth==2.17.3
google-auth-oauthlib==1.0.0
gspread==5.8.0
h11==0.14.0
idna==3.4
loguru==0.7.0
multidict==6.0.4
numpy==1.24.3
oauthlib==3.2.2
openai==0.27.5
outcome==1.2.0
packaging==23.1
pinecone-client==2.2.1
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
PyYAML==6.0
requests==2.29.0
requests-oauthlib==1.3.1
rsa==4.9
selenium==4.9.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
tqdm==4.65.0
trio==0.22.0
trio-websocket==0.10.2
typing_extensions==4.5.0
urllib3==1.26.15
webdriver-manager==3.8.6
win32-setctime==1.1.0
wsproto==1.2.0
yarl==1.9.2
